refactor: remove concept of themes and simply allow overrides in the content directory
This will lend itself better to use in a container. This is still WIP because
we need to add support for allowing users to have the Staytus web server serve
their own assets (staticly) and seperate the views a bit to allow for smaller
components to be overriden when needed.